pyproject.toml
[tool.poetry]
name = "reunion"
version = "0.1.0"
description = ""
authors = ["sd2001 <im.swarnabha2001@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.2.2"
PyJWT = "^2.6.0"
django-response = "^1.1.1"
bcrypt = "^4.0.1"
psycopg2 = "^2.9.5"
Pillow = "^9.3.0"
requests = "^2.28.1"
django-environ = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
